Refactored audio card to new pattern
