+++ packages/koenig-lexical/src/utils/audioUploadHandler.js
@@ -26,10 +26,10 @@ export const audioUploadHandler = async (files, nodeKey, editor, upload) => {
 
     await editor.update(() => {
         const node = $getNodeByKey(nodeKey);
-        node.setDuration(duration);
-        node.setSrc(fileSrc);
-        node.setMimeType(mimeType);
-        node.setTitle(title);
+        node.duration = duration;
+        node.src = fileSrc;
+        node.mimeType = mimeType;
+        node.title = title;
     });
 
     return;
