+++ packages/koenig-lexical/src/utils/thumbnailUploadHandler.js
@@ -9,14 +9,14 @@ export const thumbnailUploadHandler = async (files, nodeKey, editor, upload) => 
 
     editor.getEditorState().read(() => {
         const node = $getNodeByKey(nodeKey);
-        mediaSrc = node.getSrc();
+        mediaSrc = node.src;
     });
 
     const uploadResult = await upload(files, {formData: {url: mediaSrc}});
 
     await editor.update(() => {
         const node = $getNodeByKey(nodeKey);
-        node.setThumbnailSrc(uploadResult[0].url);
+        node.thumbnailSrc = uploadResult[0].url;
     });
 
     return;
